pyproject.toml
[tool.poetry]
name = "powerbi-ext"
version = "0.0.1"
description = "`powerbi-ext` is a Meltano utility extension."
authors = ["Leo Grosjean"]
license = "Apache 2.0"
packages = [
    { include = "powerbi_ext" },
    # uncomment this if you want to include static assets
    # { include = "files_powerbi-ext.*" }
]
# If you need to static assets with your extension, you can add them here.
# Under a directory path of "files_powerbi-ext/"
#include = [
#    "files_powerbi-ext/somefile.thing"
#]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
PyYAML = "^6.0.0"
click = "^8.1.3"
typer = "^0.6.1"
"meltano.edk" = { git = "https://github.com/meltano/edk.git", rev = "main" }
azure-identity = "^1.14.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
powerbi-ext = 'powerbi_ext.main:app'
